+++ src/cmd/cc/pickle.c
@@ -125,7 +125,7 @@ pickleinit(void)
 }
 
 void
-picklemember(Type *t, long off)
+picklemember(Type *t, int32 off)
 {
 	Sym *s, *s1;
 	static int picklecharinit = 0;
